Actualizo pom y rehago backend plato con @Model

+++ app/src/main/java/g5/app/PlatoRepository.java
@@ -3,7 +3,7 @@ package g5.app;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import java.util.*;
 
-public interface PlatoRepository extends MongoRepository<Plato,Integer>{
+public interface PlatoRepository extends MongoRepository<Plato,Long>{
 
 	Optional<Plato> findBynombre(String nombre);
